fix(v3-datetime): fold in the independent-validation findings on #878
- MonthlyMeanAccumulator accumulates its running sums in float64. A
  float32 sum drifted ~2e-4 K from the exact mean over a month of
  10-minute means, and state_dict round-trips the sums as Python floats,
  so a resumed stream was not bit-identical to an uninterrupted one. New
  regression: a float32 10-minute month is exact to 1e-12 and resuming
  through state_dict at an arbitrary seam is bit-identical.
- load_checkpoint's return value is documented precisely rather than
  changed: it is the donor's recorded elapsed_days. On a resume that equals
  the restored clock (cross-checked); with as_initial_condition=True the
  clock restarts at start_time and the value is donor provenance only,
  which initial_states.checkpoint_state relies on for its donor_days.
  Returning 0 there would break that consumer for no gain; the chunked
  runner only calls it in resume mode. The unstamped-import test now also
  asserts the reset clock and step.
- Migration guide: interval midpoints of a Model run are always whole
  seconds (averaged intervals must span an even number of seconds); the
  millisecond output unit only serves external half-second midpoints.

+++ jcm/temporal_aggregation.py
@@ -289,7 +289,14 @@ class MonthlyMeanAccumulator:
             for name in variables:
                 value = ds[name].isel(time=i, drop=True)
                 valid = value.notnull()
-                contribution = value.fillna(0) * duration_ms
+                # Accumulate in float64 whatever the field's dtype: a float32
+                # running sum drifts with the number of intervals (~2e-4 K
+                # over a month of 10-minute means), and ``state_dict`` stores
+                # the sums as Python floats, so float64 is also the only
+                # dtype in which a resumed stream is bit-identical to an
+                # uninterrupted one. The batch ``monthly_means`` reduces in
+                # float64 too.
+                contribution = value.fillna(0).astype(np.float64) * duration_ms
                 valid_duration = valid.astype(np.int64) * duration_ms
                 if name not in self._sums:
                     self._sums[name] = contribution
